- Creating the Global Action

+++ src/Controller/Admin/QuestionCrudController.php
@@ -5,14 +5,17 @@ namespace App\Controller\Admin;
 use App\EasyAdmin\VotesField;
 use App\Entity\Question;
 use App\Entity\User;
+use App\Service\CsvExporter;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\QueryBuilder;
+use EasyCorp\Bundle\EasyAdminBundle\Collection\FieldCollection;
 use EasyCorp\Bundle\EasyAdminBundle\Config\Action;
 use EasyCorp\Bundle\EasyAdminBundle\Config\Actions;
 use EasyCorp\Bundle\EasyAdminBundle\Config\Crud;
 use EasyCorp\Bundle\EasyAdminBundle\Config\Filters;
 use EasyCorp\Bundle\EasyAdminBundle\Context\AdminContext;
 use EasyCorp\Bundle\EasyAdminBundle\Controller\AbstractCrudController;
+use EasyCorp\Bundle\EasyAdminBundle\Factory\FilterFactory;
 use EasyCorp\Bundle\EasyAdminBundle\Field\AssociationField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\DateField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\Field;
@@ -20,9 +23,9 @@ use EasyCorp\Bundle\EasyAdminBundle\Field\IdField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\TextareaField;
 use EasyCorp\Bundle\EasyAdminBundle\Field\TextField;
 use EasyCorp\Bundle\EasyAdminBundle\Router\AdminUrlGenerator;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
+use Exception;
+use LogicException;
 use Symfony\Component\HttpFoundation\RedirectResponse;
-use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 
 //#[IsGranted("ROLE_SUPER_ADMIN")] // Restricting Access to an Entire Crud Section
 class QuestionCrudController extends AbstractCrudController
@@ -32,84 +35,45 @@ class QuestionCrudController extends AbstractCrudController
         return Question::class;
     }
 
-
-    public function configureFields(string $pageName): iterable
-    {
-        yield IdField::new('id')
-            ->onlyOnIndex();
-        yield Field::new('slug')
-            ->hideOnIndex()
-            ->setFormTypeOption('disabled', $pageName !== Crud::PAGE_NEW);
-        yield TextField::new('name');
-        yield AssociationField::new('topic')
-            ->autocomplete();
-        yield TextAreaField::new('question')
-            ->hideOnIndex()
-            ->setFormTypeOptions([
-                                     'row_attr' => [
-                                         'data-controller' => 'snarkdown',
-                                     ],
-                                     'attr' => [
-                                         'data-snarkdown-target' => 'input',
-                                         'data-action' => 'snarkdown#render',
-                                     ],
-                                 ])
-            ->setHelp('Preview:');
-        yield VotesField::new('votes', 'Total votes')
-            ->setTextAlign('center');
-        yield AssociationField::new('askedBy')
-            ->autocomplete()
-            ->formatValue(static function ($value,?Question $question) {
-                if (!$user = $question?->getAskedBy()) {
-                    return null;
-                }
-                return sprintf('%s&nbsp;(%s)',$user->getEmail(), $user->getQuestions()->count());
-            })
-            ->setQueryBuilder(function (QueryBuilder $queryBuilder) {
-                $queryBuilder->andWhere('entity.enabled = :enabled');
-                $queryBuilder->setParameter('enabled', true);
-            });
-        yield AssociationField::new('answers')
-            ->autocomplete()
-            ->setFormTypeOption('by_reference', false);
-        yield DateField::new('createdAt')
-            ->onlyOnIndex();
-
-        yield AssociationField::new('updatedBy')
-            ->onlyOnDetail();
-    }
-
     public function configureCrud(Crud $crud): Crud
     {
-        return parent::configureCrud($crud)
-            ->setDefaultSort([
+        return parent::configureCrud($crud)->setDefaultSort(
+            [
                 'askedBy.enabled' => 'DESC',
-                'createdAt' => 'DESC',
-                             ]);
+                'createdAt'       => 'DESC',
+            ]
+        );
     }
 
     public function configureActions(Actions $actions): Actions
     {
-        $viewAction = function () {
-            return Action::NEW('view')
-                  ->linkToUrl(function(Question $question){
-                      return $this->generateUrl('app_question_show', [
-                          'slug' => $question->getSlug()
-                      ]);
-                  })
-                  ->setIcon('fas fa-eye')
-                  ->setLabel('View on site');
-        };
+        $viewAction = function ()
+            {
+                return Action::NEW('view')->linkToUrl(
+                    function (Question $question)
+                        {
+                            return $this->generateUrl(
+                                'app_question_show',
+                                [
+                                    'slug' => $question->getSlug(),
+                                ]
+                            );
+                        }
+                )->setIcon('fas fa-eye')->setLabel('View on site');
+            };
 
-        $approveAction =  Action::new('approve')
-                ->linkToCrudAction('approve')
-                ->addCssClass('btn btn-success')
-                ->setIcon('fas fa-check-circle')
-                ->setTemplatePath('admin/approve_action.html.twig')
-                ->displayIf(static function (Question $question): bool {
+        $approveAction = Action::new('approve')->linkToCrudAction('approve')->addCssClass('btn btn-success')->setIcon(
+            'fas fa-check-circle'
+        )->setTemplatePath('admin/approve_action.html.twig')->displayIf(
+            static function (Question $question): bool
+                {
                     return !$question->getIsApproved();
-                })
-                ->displayAsButton();
+                }
+        )->displayAsButton();
+
+        $exportAction = Action::new('export')->linkToCrudAction('export')->addCssClass('btn btn-secondary')->setIcon(
+            'fas fa-download'
+        )->createAsGlobalAction();
 
         return parent::configureActions($actions)
 //            ->update(Crud::PAGE_INDEX,Action::DELETE, function (Action $action) {
@@ -123,31 +87,29 @@ class QuestionCrudController extends AbstractCrudController
             ->setPermission(Action::DELETE, 'ROLE_SUPER_ADMIN')
             ->setPermission(Action::NEW, 'ROLE_SUPER_ADMIN')
             ->setPermission(Action::BATCH_DELETE, 'ROLE_SUPER_ADMIN')
-            ->add(Action::DETAIL,$viewAction()->addCssClass('btn btn-sm btn-success'))
-            ->add(Action::INDEX, $viewAction())
-            ->add(Action::DETAIL, $approveAction);
+            ->add(Crud::PAGE_DETAIL, $viewAction()->addCssClass('btn btn-sm btn-success'))
+            ->add(Crud::PAGE_INDEX, $viewAction())
+            ->add(Crud::PAGE_DETAIL, $approveAction)
+            ->add(Crud::PAGE_INDEX, $exportAction);
     }
 
     public function configureFilters(Filters $filters): Filters
     {
-        return parent::configureFilters($filters)
-            ->add('topic')
-            ->add('askedBy')
-            ->add('createdAt')
-            ->add('votes')
-            ->add('name');
+        return parent::configureFilters($filters)->add('topic')->add('askedBy')->add('createdAt')->add('votes')->add(
+            'name'
+        );
     }
 
     public function updateEntity(EntityManagerInterface $entityManager, $entityInstance): void
     {
         //Same as create a Event Subscriber
-        if (!$entityInstance instanceof Question){
+        if (!$entityInstance instanceof Question) {
             return;
         }
         $user = $this->getUser();
 
-        if (!$user instanceof User){
-            throw new \LogicException('User should be an instance of User');
+        if (!$user instanceof User) {
+            throw new LogicException('User should be an instance of User');
         }
 
         $entityInstance->setUpdatedBy($user);
@@ -157,30 +119,84 @@ class QuestionCrudController extends AbstractCrudController
 
     public function deleteEntity(EntityManagerInterface $entityManager, $entityInstance): void
     {
-//        if ($entityInstance->getIsApproved()) {
-//            throw new \Exception('Deleting approved question is forbidden');
-//        }
+        if ($entityInstance->getIsApproved()) {
+            throw new Exception('Deleting approved question is forbidden');
+        }
 
         parent::deleteEntity($entityManager, $entityInstance);
     }
 
-
-    public function approve(AdminContext $adminContext, EntityManagerInterface $entityManager, AdminUrlGenerator $adminUrlGenerator): RedirectResponse
-    {
+    public function approve(
+        AdminContext $adminContext,
+        EntityManagerInterface $entityManager,
+        AdminUrlGenerator $adminUrlGenerator
+    ): RedirectResponse {
         $question = $adminContext->getEntity()->getInstance();
 
-        if (!$question instanceof Question){
-            throw new \LogicException('Question should be an instance of Question');
+        if (!$question instanceof Question) {
+            throw new LogicException('Question should be an instance of Question');
         }
         $question->setIsApproved(true);
         $entityManager->flush();
 
-        $targetUrl = $adminUrlGenerator
-            ->setController(self::class)
-            ->setAction(Crud::PAGE_INDEX)
-            ->setEntityId($question->getId())
-            ->generateUrl();
+        $targetUrl = $adminUrlGenerator->setController(self::class)->setAction(Crud::PAGE_INDEX)->setEntityId(
+            $question->getId()
+        )->generateUrl();
 
         return $this->redirect($targetUrl);
     }
+
+    public function export(AdminContext $context, CsvExporter $csvExporter)
+    {
+        $fields = FieldCollection::new($this->configureFields(Crud::PAGE_INDEX));
+        $filters = $this->container->get(FilterFactory::class)->create(
+            $context->getCrud()->getFiltersConfig(),
+            $fields,
+            $context->getEntity()
+        );
+        $queryBuilder = $this->createIndexQueryBuilder($context->getSearch(), $context->getEntity(), $fields, $filters);
+
+        return $csvExporter->createResponseFromQueryBuilder($queryBuilder, $fields, 'questions.csv');
+    }
+
+    public function configureFields(string $pageName): iterable
+    {
+        yield IdField::new('id')->onlyOnIndex();
+        yield Field::new('slug')->hideOnIndex()->setFormTypeOption('disabled', $pageName !== Crud::PAGE_NEW);
+        yield TextField::new('name');
+        yield AssociationField::new('topic')->autocomplete();
+        yield TextAreaField::new('question')->hideOnIndex()->setFormTypeOptions(
+            [
+                'row_attr' => [
+                    'data-controller' => 'snarkdown',
+                ],
+                'attr'     => [
+                    'data-snarkdown-target' => 'input',
+                    'data-action'           => 'snarkdown#render',
+                ],
+            ]
+        )->setHelp('Preview:');
+        yield VotesField::new('votes', 'Total votes')->setTextAlign('center');
+        yield AssociationField::new('askedBy')->autocomplete()->formatValue(
+            static function ($value, ?Question $question)
+                {
+                    if (!$user = $question?->getAskedBy()) {
+                        return null;
+                    }
+
+                    return sprintf('%s&nbsp;(%s)', $user->getEmail(), $user->getQuestions()->count());
+                }
+        )->setQueryBuilder(
+            function (QueryBuilder $queryBuilder)
+                {
+                    $queryBuilder->andWhere('entity.enabled = :enabled');
+                    $queryBuilder->setParameter('enabled', true);
+                }
+        );
+        yield AssociationField::new('answers')->autocomplete()->setFormTypeOption('by_reference', false);
+        yield DateField::new('createdAt')->onlyOnIndex();
+
+        yield AssociationField::new('updatedBy')->onlyOnDetail();
+    }
+
 }
